refactor: 체스 보드로 책임 이동
